feat: migrate mock data and edit README
