+++ no-tang-doc-web/vite.config.ts
@@ -45,7 +45,6 @@ export default defineConfig({
       '@radix-ui/react-aspect-ratio@1.1.2': '@radix-ui/react-aspect-ratio',
       '@radix-ui/react-alert-dialog@1.1.6': '@radix-ui/react-alert-dialog',
       '@radix-ui/react-accordion@1.2.3': '@radix-ui/react-accordion',
-      '@jsr/supabase__supabase-js@2.49.8': '@jsr/supabase__supabase-js',
       '@': path.resolve(__dirname, './src'),
     },
   },
